requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
requests==2.32.0
pytest==7.4.2
httpx==0.27.0
sentence_transformers==2.3.0
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv
google-generativeai